[GH actions] Try to remove tags from create_release.
This GitHub Actions task has shitty problem that it creates a tag that
conflicts the branch name and brings in messes later. It is simply not
acceptable by design and they should fix the problem.

Fortunately it was already reported (and already 10 days ago, meaning that
they don't plan to fix the issue?? It is surprisingly the official repo)
https://github.com/actions/create-release/issues/13